# docs/swagger.yaml
basePath: /
definitions:
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.ApiListData:
    properties:
      comment:
        example: Products needed for the party
        type: string
      name:
        example: List of products
        type: string
    type: object
  models.ApiMessage:
    properties:
      message:
        type: string
    type: object
  models.ApiShowLists:
    properties:
      lists:
        items:
          $ref: '#/definitions/models.ListsData'
        type: array
    type: object
...
  models.ListEditData:
    properties:
      comment:
        example: Products needed for the party
        type: string
      id:
        example: 1023456789
        type: integer
      index:
        example: 1
        type: integer
      name:
        example: New list of products
        type: string
    type: object
  models.ListsData:
    properties:
      comment:
        example: Products needed for the party
        type: string
      id:
        example: 1023456789
        type: integer
      index:
        example: 0
        type: integer
      name:
        example: List of products
        type: string
    type: object
...
  models.UserData:
    properties:
      email:
        example: nktkln@example.com
        type: string
      name:
        example: NKTKLN
        type: string
      password:
        example: StRon9Pa$$w0rd
        type: string
      username:
        example: nktkln
        type: string
    type: object
...
  models.UserLoginData:
    properties:
      email:
        example: nktkln@example.com
        type: string
      password:
        example: StRon9Pa$$w0rd
        type: string
    type: object
...
  models.UserTokens:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzMDE5NDgxODQiLCJleHAiOjE2NTU0MTc4MDF9.ZqPxmPK3qV3VrT4D0wbwN2tMGzAhaH5kQnqr8iePTZA
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzMDE5NDgxODQiLCJleHAiOjE2NTU0MjI1NTh9.S1ypUshOnPB66VJV6RL42cdgYbV8LaGk7zYgL5JlsYg
        type: string
    type: object
...
info:
  contact:
    email: nktkln@nktkln.com
    name: API Support
    url: https://nktkln.com
  description: some description
  license:
    name: MIT
    url: http://github.com/
  title: ToDo API
  version: "1.0"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: LoginData
        required: true
        schema:
          $ref: '#/definitions/models.UserLoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Sign in to your account
      tags:
      - Authorization
  /auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: User data
        in: body
        name: UserData
        required: true
        schema:
          $ref: '#/definitions/models.UserData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Register a user
      tags:
      - Authorization
  /auth/verify:
    get:
      consumes:
      - application/json
      parameters:
      - description: Verification key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Confirm the new user's email
      tags:
      - Authorization
  /todo/list/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: List data
        in: body
        name: ListData
        required: true
        schema:
          $ref: '#/definitions/models.ApiListData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Create list
      tags:
      - Working with lists
  /todo/list/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: The id of the list to be deleted
        in: query
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Delete list
      tags:
      - Working with lists
  /todo/list/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: List data
        in: body
        name: ListData
        required: true
        schema:
          $ref: '#/definitions/models.ListEditData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Edit list
      tags:
      - Working with lists
  /todo/list/show:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiShowLists'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - token: []
      summary: Shows all lists created by the user
      tags:
      - Working with lists
...
securityDefinitions:
  token:
    in: header
    name: token
    type: apiKey
swagger: "2.0"